+++ python/python/tests/test_transaction.py
@@ -56,9 +56,12 @@ async def transaction_test(
     transaction.exists([key])
     args.append(1)
 
-    transaction.delete([key])
+    transaction.rename(key, key2)
+    args.append(OK)
+
+    transaction.delete([key2])
     args.append(1)
-    transaction.get(key)
+    transaction.get(key2)
     args.append(None)
 
     transaction.mset({key: value, key2: value2})